refactor bit-use and bit-merge, remove unused flags from bit-merge
