+++ src/consumer/versions-ops/merge-version/three-way-merge.js
@@ -1,17 +1,17 @@
 // @flow
 import R from 'ramda';
-import Component from '../component';
-import { Version } from '../../scope/models';
-import { Consumer } from '..';
-import { sha1 } from '../../utils';
-import { SourceFile } from '../component/sources';
-import { Tmp } from '../../scope/repositories';
-import mergeFiles from './merge-files';
-import type { MergeFileResult, MergeFileParams } from './merge-files';
-import type { PathOsBased } from '../../utils/path';
-import type { SourceFileModel } from '../../scope/models/version';
+import Component from '../../component';
+import { Version } from '../../../scope/models';
+import { Consumer } from '../..';
+import { sha1 } from '../../../utils';
+import { SourceFile } from '../../component/sources';
+import { Tmp } from '../../../scope/repositories';
+import mergeFiles from '../../../utils/merge-files';
+import type { MergeFileResult, MergeFileParams } from '../../../utils/merge-files';
+import type { PathOsBased } from '../../../utils/path';
+import type { SourceFileModel } from '../../../scope/models/version';
 
-export type MergeResults = {
+export type MergeResultsThreeWay = {
   addFiles: Array<{
     filePath: string,
     fsFile: SourceFile
@@ -35,25 +35,6 @@ export type MergeResults = {
   hasConflicts: boolean
 };
 
-export type MergeResultsTwoWay = {
-  addFiles: Array<{
-    filePath: string,
-    otherFile: SourceFileModel
-  }>,
-  modifiedFiles: Array<{
-    filePath: string,
-    otherFile: SourceFileModel,
-    currentFile: SourceFile,
-    output: ?string,
-    conflict: ?string
-  }>,
-  unModifiedFiles: Array<{
-    filePath: string,
-    currentFile: SourceFile
-  }>,
-  hasConflicts: boolean
-};
-
 /**
  * it's easier to understand with an example.
  * a component bar/foo has two versions: 0.0.1, 0.0.2. Also, the component was modified locally.
@@ -73,7 +54,7 @@ export type MergeResultsTwoWay = {
  * base-file    => bar/foo@0.0.2
  * other-file   => bar/foo@0.0.2 + modification
  */
-export async function threeWayMergeVersions({
+export default (async function threeWayMergeVersions({
   consumer,
   otherComponent,
   otherVersion,
@@ -87,7 +68,7 @@ export async function threeWayMergeVersions({
   currentComponent: Version,
   currentVersion: string,
   baseComponent: Version
-}): Promise<MergeResults> {
+}): Promise<MergeResultsThreeWay> {
   // baseFiles and currentFiles come from the model, therefore their paths include the
   // sharedOriginallyDir. fsFiles come from the Fs, therefore their paths don't include the
   // sharedOriginallyDir.
@@ -158,72 +139,11 @@ export async function threeWayMergeVersions({
   });
 
   return results;
-}
-
-export async function twoWayMergeVersions({
-  consumer,
-  otherComponent,
-  otherVersion,
-  currentComponent,
-  currentVersion
-}: {
-  consumer: Consumer,
-  otherComponent: Version,
-  otherVersion: string,
-  currentComponent: Component,
-  currentVersion: string
-}): Promise<MergeResultsTwoWay> {
-  const otherFiles: SourceFileModel[] = otherComponent.files;
-  const currentFiles: SourceFile[] = currentComponent.cloneFilesWithSharedDir();
-  const results: MergeResultsTwoWay = {
-    addFiles: [],
-    modifiedFiles: [],
-    unModifiedFiles: [],
-    overrideFiles: [],
-    hasConflicts: false
-  };
-  const getFileResult = (otherFile: SourceFileModel, currentFile?: SourceFile) => {
-    const filePath = otherFile.relativePath;
-    if (!currentFile) {
-      results.addFiles.push({ filePath, otherFile });
-      return;
-    }
-    const otherFileHash = otherFile.file.hash;
-    const currentFileHash = sha1(currentFile.contents);
-    if (otherFileHash === currentFileHash) {
-      results.unModifiedFiles.push({ filePath, currentFile });
-      return;
-    }
-    // it was changed in both, there is a chance for conflict
-    currentFile.version = currentVersion;
-    // $FlowFixMe it's a hack to pass the data, version is not a valid attribute.
-    otherFile.version = otherVersion;
-    results.modifiedFiles.push({ filePath, currentFile, otherFile, output: null, conflict: null });
-  };
-
-  otherFiles.forEach((otherFile) => {
-    const relativePath = otherFile.relativePath;
-    const currentFile = currentFiles.find(file => file.relative === relativePath);
-    getFileResult(otherFile, currentFile);
-  });
-
-  if (R.isEmpty(results.modifiedFiles)) return results;
-
-  const conflictResults = await getConflictResultsTwoWay(consumer, results.modifiedFiles);
-  conflictResults.forEach((conflictResult) => {
-    const modifiedFile = results.modifiedFiles.find(file => file.filePath === conflictResult.filePath);
-    if (!modifiedFile) throw new Error(`unable to find ${conflictResult.filePath} in modified files array`);
-    modifiedFile.output = conflictResult.output;
-    modifiedFile.conflict = conflictResult.conflict;
-    if (conflictResult.conflict) results.hasConflicts = true;
-  });
-
-  return results;
-}
+});
 
 async function getConflictResults(
   consumer: Consumer,
-  modifiedFiles: $PropertyType<MergeResults, 'modifiedFiles'>
+  modifiedFiles: $PropertyType<MergeResultsThreeWay, 'modifiedFiles'>
 ): Promise<MergeFileResult[]> {
   const tmp = new Tmp(consumer.scope);
   const conflictResultsP = modifiedFiles.map(async (modifiedFile) => {
@@ -252,51 +172,6 @@ async function getConflictResults(
       otherFile: {
         label: `${modifiedFile.fsFile.version} modified`,
         path: fsFilePath
-      }
-    };
-    return mergeFiles(mergeFilesParams);
-  });
-  try {
-    const conflictResults = await Promise.all(conflictResultsP);
-    await tmp.clear();
-    return conflictResults;
-  } catch (err) {
-    await tmp.clear();
-    throw err;
-  }
-}
-
-async function getConflictResultsTwoWay(
-  consumer: Consumer,
-  modifiedFiles: $PropertyType<MergeResultsTwoWay, 'modifiedFiles'>
-): Promise<MergeFileResult[]> {
-  const tmp = new Tmp(consumer.scope);
-  const conflictResultsP = modifiedFiles.map(async (modifiedFile) => {
-    const currentFilePathP = tmp.save(modifiedFile.currentFile.contents);
-    const writeFile = async (file: SourceFileModel): Promise<PathOsBased> => {
-      const content = await file.file.load(consumer.scope.objects);
-      return tmp.save(content);
-    };
-    const baseFilePathP = tmp.save('');
-    const otherFilePathP = writeFile(modifiedFile.otherFile);
-    const [otherFilePath, baseFilePath, currentFilePath] = await Promise.all([
-      otherFilePathP,
-      baseFilePathP,
-      currentFilePathP
-    ]);
-    const mergeFilesParams: MergeFileParams = {
-      filePath: modifiedFile.filePath,
-      currentFile: {
-        label: modifiedFile.currentFile.version,
-        path: currentFilePath
-      },
-      baseFile: {
-        path: baseFilePath
-      },
-      otherFile: {
-        // $FlowFixMe
-        label: modifiedFile.otherFile.version,
-        path: otherFilePath
       }
     };
     return mergeFiles(mergeFilesParams);
